Grid/Categories: wire up nested routes, support parseable values:
Can now visit /grid/column/value routes and see a proper list. And the
/grid/column route supports comma'ed values like Countries and
Neighborhoods.
